©®ŋ®ŧħ←ø°đßł©─ŋ”®°?®ßđð©đßð“
