+++ config.py
@@ -1,0 +1,20 @@
+import pygame
+
+# Dimensões da tela
+WIDTH, HEIGHT = 1000, 600
+
+# Cores
+WHITE = (255, 255, 255)
+GRAY = (100, 100, 100)
+RED = (200, 50, 50)
+BLUE = (50, 100, 200)
+BLACK = (0, 0, 0)
+GREEN = (50, 200, 50)
+YELLOW = (255, 255, 0)
+ORANGE = (255, 165, 0)
+PURPLE = (128, 0, 128)
+
+# Inicialização das fontes
+pygame.font.init()
+font = pygame.font.SysFont("Arial", 20)
+large_font = pygame.font.SysFont("Arial", 30)
