+++ requirements.txt
@@ -1,0 +1,2 @@
+pygame==2.5.2
+numpy==1.26.3
